fetch/metadata/generated/fetch-sec-fetch-site.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection from HTTP</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  promise_test(() => {
    const key = '{{uuid()}}';

    return fetch(
        makeRequestURL(key, ['sameOrigin', 'httpOrigin']),
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'HTTPS downgrade (header not sent)');

  promise_test(() => {
    const key = '{{uuid()}}';

    return fetch(
        makeRequestURL(key, ['httpOrigin', 'sameOrigin']),
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS upgrade');

  promise_test(() => {
    const key = '{{uuid()}}';

    return fetch(
        makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin']),
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'HTTPS downgrade-upgrade');
  </script>
  </body>
</html>
